+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.0.1] - 2024-06-27
+## [0.0.1] - 2024-06-28
 
 ### Changed
 
  - updated verified and finalized states in negotiation module
  - separated consumer and provider callback addresses in negotiation module
+
+## [0.0.1] - 2024-06-25
+
+### Added
+
+ - DataTransfer Consumer callback controller and junit tests
 
 ## [0.0.1] - 2024-06-21